refactor(rippling): extract digest mock-up + post-detail modal
Second step of breaking RipplingExplorer.vue down.  All the modal markup,
CSS and rendering logic for the three modal flavours (full digest mock-up,
single-post deep-dive, same-location cluster) moves into the new
RipplingDigestModal.vue component.

The component is self-contained: it owns the DOM (including
component-scoped CSS), tracks its own open/closed state, and exposes
three methods via defineExpose — openDigest(ranked, lat, lng),
openPost(post, rank, lat, lng), openCluster(posts, lat, lng).  Member
location is passed in per call so the modal stays stateless about it.

RipplingExplorer.vue loses ~290 lines of inline modal handling (markup,
CSS, three large render functions, the close handlers) and gains a
single <RipplingDigestModal ref="digestModal" /> mount plus three
one-line shims.  Zero behaviour change.

File now ~3000 lines, down from ~3300 before this and the previous
helper-extraction commit.
